+++ R/class-chart-sheet.R
@@ -65,13 +65,14 @@ wbChartSheet <- R6::R6Class(
       self$sheetViews  <- character()
       self$pageMargins <- '<pageMargins left="0.7" right="0.7" top="0.75" bottom="0.75" header="0.3" footer="0.3"/>'
       self$drawing     <- '<drawing r:id=\"rId1\"/>'
-      self$relships              <- list(
+      self$relships    <- list(
         comments         = integer(),
         drawing          = integer(),
         pivotTable       = integer(),
         slicer           = integer(),
         table            = integer(),
         threadedComments = integer(),
+        timeline         = integer(),
         vmlDrawing       = integer()
       )